refactor: extract shared GitHubLink component and tweak accent color
Create a reusable GitHubLink component used in both the nav header
and hero. Adjust accent color to be slightly darker in light mode
and lighter in dark mode. Clean up redundant icon classes in SignInButton.
